Partition apiUrls into smaller files.
