+++ src/shared/config/apiUrls/apiPrefix.ts
@@ -1,0 +1,2 @@
+// Injected by webpack. Override for adhoc local testing.
+export const apiPrefix = API_PREFIX;
