+++ src/shared/config/apiUrls/apiUrls.ts
@@ -1,0 +1,15 @@
+import * as configure from './configure';
+import * as entry from './entry';
+import * as proteinsApi from './proteinsApi';
+import * as search from './search';
+import * as results from './results';
+import * as suggester from './suggester';
+
+export default {
+  configure,
+  entry,
+  proteinsApi,
+  results,
+  search,
+  suggester,
+};
