+++ src/shared/config/apiUrls/suggester.ts
@@ -1,0 +1,12 @@
+import joinUrl from 'url-join';
+
+import { apiPrefix } from './apiPrefix';
+
+export const taxonomySuggester = 'suggester?dict=taxonomy&query=?';
+
+export const organismSuggester = 'suggester?dict=organism&query=?';
+
+const RE_QUERY = /\?$/;
+
+export const getSuggesterUrl = (url: string, value: string) =>
+  joinUrl(apiPrefix, url.replace(RE_QUERY, value));
